renaming the home.jsx file to index.jsx
